+++ gems/aws-sdk-ec2/lib/aws-sdk-ec2/client.rb
@@ -58776,7 +58776,7 @@ module Aws::EC2
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-ec2'
-      context[:gem_version] = '1.438.0'
+      context[:gem_version] = '1.439.0'
       Seahorse::Client::Request.new(handlers, context)
     end