+++ gems/aws-sdk-wafv2/lib/aws-sdk-wafv2/types.rb
@@ -6161,6 +6161,20 @@ module Aws::WAFV2
     #     method, city pair.
     #   @return [Integer]
     #
+    # @!attribute [rw] evaluation_window_sec
+    #   The amount of time, in seconds, that WAF should include in its
+    #   request counts, looking back from the current time. For example, for
+    #   a setting of 120, when WAF checks the rate, it counts the requests
+    #   for the 2 minutes immediately preceding the current time. Valid
+    #   settings are 60, 120, 300, and 600.
+    #
+    #   This setting doesn't determine how often WAF checks the rate, but
+    #   how far back it looks each time it checks. WAF checks the rate about
+    #   every 10 seconds.
+    #
+    #   Default: `300` (5 minutes)
+    #   @return [Integer]
+    #
     # @!attribute [rw] aggregate_key_type
     #   Setting that indicates how to aggregate the request counts.
     #
@@ -6240,6 +6254,7 @@ module Aws::WAFV2
     #
     class RateBasedStatement < Struct.new(
       :limit,
+      :evaluation_window_sec,
       :aggregate_key_type,
       :scope_down_statement,
       :forwarded_ip_config,